extendedWriteMultipleBlocks worked
As the API was not available in the flutter-nfc-manager plugin, had to import the plugin in the local project structure (instead of its link to pub.dev) and then put the API in it.
Had to change iOS support from 13 to 14 for this API's sake.
With the API working, 4 blocks can be written at a time and hence the timing is reduced from 5.2s to 2.8s for 1200 bytes.